Delete importers
Removing code that we don't need, to reduce the amount of code that needs to be updated
