+++ src/main/java/bq_standard/core/proxies/ClientProxy.java
@@ -3,10 +3,7 @@ package bq_standard.core.proxies;
 import betterquesting.api.api.ApiReference;
 import betterquesting.api.api.QuestingAPI;
 import bq_standard.client.theme.BQSTextures;
-import bq_standard.importers.NativeFileImporter;
-import bq_standard.importers.ftbq.FTBQQuestImporter;
-import bq_standard.importers.hqm.HQMBagImporter;
-import bq_standard.importers.hqm.HQMQuestImporter;
+import bq_standard.importers.DummyImporter;
 import bq_standard.integration.nei.IMCForNEI;
 
 public class ClientProxy extends CommonProxy {
@@ -26,12 +23,7 @@ public class ClientProxy extends CommonProxy {
     public void registerExpansion() {
         super.registerExpansion();
 
-        QuestingAPI.getAPI(ApiReference.IMPORT_REG).registerImporter(NativeFileImporter.INSTANCE);
-
-        QuestingAPI.getAPI(ApiReference.IMPORT_REG).registerImporter(HQMQuestImporter.INSTANCE);
-        QuestingAPI.getAPI(ApiReference.IMPORT_REG).registerImporter(HQMBagImporter.INSTANCE);
-
-        QuestingAPI.getAPI(ApiReference.IMPORT_REG).registerImporter(FTBQQuestImporter.INSTANCE);
+        QuestingAPI.getAPI(ApiReference.IMPORT_REG).registerImporter(DummyImporter.INSTANCE);
 
         BQSTextures.registerTextures();
     }
