--- src/main/java/bq_standard/importers/hqm/converters/items/HQMItem.java
@@ -1,8 +1,0 @@
-package bq_standard.importers.hqm.converters.items;
-
-import betterquesting.api.utils.BigItemStack;
-import net.minecraft.nbt.NBTTagCompound;
-
-public interface HQMItem {
-    BigItemStack convertItem(int damage, int amount, NBTTagCompound tags);
-}
