+++ src/main/resources/assets/bq_standard/lang/zh_CN.lang
@@ -54,18 +54,8 @@ bq_standard.gui.experience=经验
 bq_standard.gui.tame=驯养 %s
 bq_standard.gui.questcompletion=任务完成
 
-bq_standard.importer.hqm_bag.name=HQM奖励袋导入
-bq_standard.importer.hqm_bag.desc=导入以json格式导出的HQM奖励袋.
-bq_standard.importer.hqm_quest.name=HQM任务导入
-bq_standard.importer.hqm_quest.desc=导入以json格式导出的HQM任务. 由于HQM任务在导出后被索引的奇怪特性, 可能会在具有重复名称或条件解锁的任务之间发生错误.
-bq_standard.importer.nat_file.name=本地文件导入
-bq_standard.importer.nat_file.desc=导入本地任务数据并将其附加到现有任务中.
-bq_standard.importer.nat_url.name=本地Url导入
-bq_standard.importer.nat_url.desc=从互联网导入本地任务数据并将其附加到现有任务中.
-bq_standard.importer.ftbq_quest.name=FTB任务导入
-bq_standard.importer.ftbq_quest.desc=从FTB Quests mod的本地NBT文件导入章节和任务
-bq_standard.importer.adv_json.name=进阶导入 (JSON)
-bq_standard.importer.adv_json.desc=导入json格式的原版进阶(advancements).%n%n警告: 必须在默认文件地址才能成功导入!
+bq_standard.importer.dummy.name=Dummy Importer
+bq_standard.importer.dummy.desc=Does nothing but throw UnsupportedOperationException.
 
 item.bq_standard.loot_chest.name=奖励箱
